+++ lib/view/component/detail/drawer_tile.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import '../../view/theme/app_theme.dart';
+import '../../theme/app_theme.dart';
 
 class DrawerTile extends HookConsumerWidget {
   const DrawerTile({
