+++ lib/view/component/detail/graph_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import '../../provider/general_provider.dart';
-import '../../view/theme/app_theme.dart';
+import '../../../provider/general_provider.dart';
+import '../../theme/app_theme.dart';
 
 class GraphBar extends HookConsumerWidget {
   const GraphBar({
