+++ lib/view/component/home/custom_painter.dart
