+++ lib/view/component/home/home_details_tile.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../provider/general_provider.dart';
-import '../../view/theme/app_theme.dart';
+import '../../../provider/general_provider.dart';
+import '../../theme/app_theme.dart';
 
 class HomeDetailsTile extends HookConsumerWidget {
   const HomeDetailsTile({
