+++ lib/view/component/large_text_field.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../view/theme/app_theme.dart';
+import '../theme/app_theme.dart';
 
-class NormalTextField extends HookConsumerWidget {
-  NormalTextField({
+class LargeTextField extends HookConsumerWidget {
+  LargeTextField({
     super.key,
     required this.topTitle,
     required this.bottomTitle,
@@ -31,26 +31,25 @@ class NormalTextField extends HookConsumerWidget {
         const SizedBox(height: 10),
         Text(topTitle, style: theme.textTheme.fs16),
         Container(
-          padding: const EdgeInsets.symmetric(horizontal: 10),
+          padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 5),
           width: double.infinity,
-          height: 45,
+          height: 90,
           decoration: BoxDecoration(
             border: Border.all(),
             borderRadius: BorderRadius.circular(10),
           ),
-          child: Center(
-            child: TextFormField(
-              controller: controller,
-              keyboardType: inputType,
-              inputFormatters: formatter ?? [],
-              decoration: InputDecoration(
-                hintText: hintText,
-                hintStyle: theme.textTheme.fs19,
-                contentPadding: EdgeInsets.zero,
-                isCollapsed: true,
-                focusedBorder: InputBorder.none,
-                border: InputBorder.none,
-              ),
+          child: TextFormField(
+            controller: controller,
+            keyboardType: inputType,
+            inputFormatters: formatter ?? [],
+            maxLines: 3,
+            decoration: InputDecoration(
+              hintText: hintText,
+              hintStyle: theme.textTheme.fs19,
+              contentPadding: EdgeInsets.zero,
+              isCollapsed: true,
+              focusedBorder: InputBorder.none,
+              border: InputBorder.none,
             ),
           ),
         ),
