+++ lib/view/component/setting/friend_tile_with_radio.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../provider/general_provider.dart';
-import '../../view/theme/app_theme.dart';
+import '../../../provider/general_provider.dart';
+import '../../theme/app_theme.dart';
 
 class FriendTileWithRadio extends HookConsumerWidget {
   const FriendTileWithRadio({
