+++ lib/view/pages/home/add_project/target_add_member.dart
@@ -2,7 +2,7 @@ import 'package:account_book_app/constant/hex_color.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import '../../../../component/setting/friend_tile_with_radio.dart';
+import '../../../component/setting/friend_tile_with_radio.dart';
 import '../../../../provider/general_provider.dart';
 import '../../../../provider/route/routes.dart';
 import '../../../theme/app_theme.dart';
