+++ lib/view/pages/home/detail/details.dart
@@ -5,7 +5,7 @@ import 'package:account_book_app/view/pages/home/detail/saving_add.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import '../../../../component/home/home_details_tile.dart';
+import '../../../component/home/home_details_tile.dart';
 import '../../../../provider/general_provider.dart';
 import '../../../theme/app_theme.dart';
 import 'my_drawer.dart';
