+++ lib/view/pages/home/detail/panels/page_view_left.dart
@@ -1,4 +1,4 @@
-import 'package:account_book_app/component/detail/graph_bar.dart';
+import 'package:account_book_app/view/component/detail/graph_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
